Number ASCII Art Generator!!
ASCII Art Number Generator in C

This project generates ASCII art representations of numbers from 1 to 9 using the C programming language. In the future, more numbers and possibly text characters will be added to expand the functionality. The project demonstrates basic C concepts and helps in understanding loops and conditional structures in a fun, visual way